IRIS-3881 | moved RegisterArea and LoginArea to RegisterPage and SinInPage, consecutively
